Add Alexa Skill Information and Updates to Ruby Version Used in CI Pipeline and Dependencies (#57)
This commit adds information about the Panama City Beach Flags Alexa Skill to the Sea Florida Change website. It updates the version of Ruby used in the CI pipeline. It also resolves the following vulnerabilities detected by Snyk that are related to the Nokogiri dependency:

https://snyk.io/vuln/SNYK-RUBY-NOKOGIRI-2620374
https://snyk.io/vuln/SNYK-RUBY-NOKOGIRI-2630623
https://snyk.io/vuln/SNYK-RUBY-NOKOGIRI-2630898

Finally, minor updates to other dependencies were made.