Update Google Analytics Configuration and CI Pipeline (#70)
This commit updates the Google Analytics configuration used for the Sea Florida Change website. Instead of relying on a Jekyll variable, the tag is now hardcoded into the default layout. It also updates the version of the Jekyll Deploy Test Action used as part of the CI pipeline for the Sea Florida Change website.